moved the prayer preview card to the root level prayer folder
