+++ packages/newspringchurchapp/src/prayer/PrayerPreviewCard/PrayerPreviewCard.tests.js
