+++ packages/newspringchurchapp/src/prayer/PrayerPreviewCard/index.js
